Misc polishing, up to release candidate 2
* combines JeffNet and Tyler branches, both working here and now
* starts adding interfaces for things that only ECF / tyler have
  (just codes for now)
